requirements.txt
disnake>=2.8.0
toml>=0.10.2
gitpython
